// src/webApp.types.ts
export type WebApplication = {
  id: number;
  server_user_id: number;
  name: string;
  rootPath: string;
  publicPath: string;
  phpVersion: string;
  stack: string;
  stackMode: string;
  type: string;
  defaultApp: false;
  alias?: string;
  pullKey1?: string;
  pullKey2?: string;
  created_at: string;
};

export type CreateWebAppInput = {
  name: string;
  domainName: string;
  user: number;
  publicPath?: string;
  phpVersion: string;
  stack: "hybrid" | "nativenginx" | "customnginx";
  stackMode: "production" | "development";
  clickjackingProtection: boolean;
  xssProtection: boolean;
  mimeSniffingProtection: boolean;
  processManager: "dynamic" | "ondemand" | "static";
  processManagerStartServers?: number;
  processManagerMinSpareServers?: number;
  processManagerMaxSpareServers?: number;
  processManagerMaxChildren: number;
  processManagerMaxRequests: number;
  openBasedir?: string;
  timezone: TimeZone;
  disableFunctions?: string;
  maxExecutionTime: number;
  maxInputTime: number;
  maxInputVars: number;
  memoryLimit: number;
  postMaxSize: number;
  uploadMaxFilesize: number;
  sessionGcMaxlifetime: number;
  allowUrlFopen: boolean;
};

export type WebAppResponse = {
  data?: WebApplication[];
  meta: Meta;
};

export type CloneGitRepoInput = {
  serverId: number;
  webAppId: number;
  provider: "custom" | "bitbucket" | "github" | "gitlab" | "selfhostedgitlab";
  repository: string;
  branch: string;
  gitUser?: string;
  gitHost?: string;
};

export type GitObject = {
  id: number;
  provider: string;
  gitHost?: string;
  gitUser?: string;
  branch: string;
  repositoryData: {
    url: string;
    repo: string;
  };
  atomic: boolean;
  atomic_project_id?: string;
  autoDeploy: boolean;
  deployScript?: string;
  created_at: string;
};

export type InstallPHPScriptInput = {
  serverId: number;
  webAppId: number;
  name: PHPScript;
};

export type PHPResponse = {
  id: number;
  name: string;
  realName: string;
  created_at: string;
};

export type AddDomainNameInput = {
  serverId: number;
  webAppId: number;
  installerId: number;
  name: string;
  www?: boolean;
  routeWww?: boolean;
  type?: "alias" | "primary" | "redirect";
};

export type DomainName = {
  id: number;
  name: string;
  type: string;
  www: number;
  redirection: string;
  wildcard: number;
  dns_integration: string;
  status?: string;
  created_at: string;
};

export type DomainNamesReponse = {
  data?: DomainName[];
  meta: Meta;
};

export type PHPScript =
  | "concrete5"
  | "drupal"
  | "grav"
  | "gravadmin"
  | "joomla"
  | "myBB"
  | "phpBB"
  | "phpMyAdmin"
  | "piwik"
  | "prestaShop"
  | "wordpress";